Uploaded Tendable_Project.txt file
Its Selenium based project written in Java wherein i have written the script to Confirm accessibility of the top-level menus , Verifying that the "Request a Demo" button is present and active , locating the element of navigation to Marketing section and filling up form. And written the script to confirm the error message's appearance.